Move local networking errors somewhere nicer
