+++ src/transport/errors/UnexpectedResponseError.ts
@@ -1,0 +1,6 @@
+export class UnexpectedResponseError extends TypeError {
+	constructor(message: string) {
+		super(`Server response was unexpected: ${message}`);
+		this.name = "UnexpectedResponseError";
+	}
+}
