+++ src/transport/onSnapshot.ts
@@ -3,7 +3,7 @@ import type { Query } from "./db";
 import { AccountableError } from "./AccountableError.js";
 import { DocumentReference, CollectionReference } from "./db.js";
 import { isRawServerResponse } from "./schemas";
-import { UnexpectedResponseError } from "./networking.js";
+import { UnexpectedResponseError } from "./errors/index.js";
 import { UnreachableError } from "./UnreachableError.js";
 import isArray from "lodash/isArray";
 import isString from "lodash/isString";
